Version 0.8.5 adding keys to skipped attributes adding validation to draft object (#4)
* Removing Annoying Draftsman Things

A lot about this gem annoys me, but with this commit pushed I will be as calm as a cucumber, as trainquil as still lake
Removed code that checked if the draft has returned to it's original state, if it had, delete it.
This just confused our system and would cause the content to enter an un-approvable state. Also this logic seems handwavy, why magically delete the draft?
It makes sense in a way, but for our use case it does not make sense.
Removed logic around create drafts.
Again this makes very little sense to me, if the draft has an event of 'create' we update the parent object. Why?
Why not just keep updating the draft?
I guess if it's never been published then you want both to stay in sync? It doesn't make much sense to me.

Signed-off-by: Evan <evan@spectrum.md>

* Bump Version

Signed-off-by: Evan <evan@spectrum.md>

* Removed Tests

These were not actually testing much with the create draft change
They were causing a failure on push b/c of rounding in the DB

Signed-off-by: Evan <evan@spectrum.md>

* Updated README

Signed-off-by: Evan <evan@spectrum.md>

* Skip draft_id and id by default

Both of these keys are really dangerous if they end up in the draft
When duplicating an item with a draft these have a tendency to sneak into the draft object
Explicitly adding them to the skipped attributes will fix this
There's an ignore option that seems to do something similar, but it does not look like it is used really...

Signed-off-by: Evan <evan@spectrum.md>

* Added Draft Object Validator Class

Added Draft Object Validator Class
This Validates the draft to make sure that none of the skipped attributes exist in the draft object
If it does it flags the draft as being invalid

Signed-off-by: Evan <evan@spectrum.md>

* Fixing Test Scripts That Checked For ID

Id is now auto skipped so removed this from test scripts

Signed-off-by: Evan <evan@spectrum.md>

* Test That Skipper Does Not Have id in Object

Signed-off-by: Evan <evan@spectrum.md>

* Replace empty? with any? on draftsman validator

Signed-off-by: Evan <evan@spectrum.md>